requirements.txt
# torch always from the local custom ROCm 7.11 wheel:
torch @ file:///opt/rocm/wheels/pytorch_rocm711/torch-2.11.0a0+git3b6829f-cp312-cp312-linux_x86_64.whl

# deps from the internet are fine:
openai-whisper
numpy

